CM-208: refactored converter
