+++ calcrule_social_protection/converters/beneficiary/beneficiary_to_bill_item.py
@@ -1,37 +1,8 @@
-from django.contrib.contenttypes.models import ContentType
+from calcrule_social_protection.converters.base import BaseBenefitPackageStrategyItemConverter
 
 
-class BeneficiaryToBillItemConverter:
-
-    @classmethod
-    def to_bill_item_obj(cls, payment_plan, beneficiary, amount):
-        bill_line_item = {}
-        cls._build_line_fk(bill_line_item, beneficiary)
-        cls._build_code(bill_line_item, beneficiary)
-        cls._build_date_dates(bill_line_item, payment_plan)
-        cls._build_price(bill_line_item, amount)
-        cls._build_quantity(bill_line_item)
-        return bill_line_item
-    
-    @classmethod
-    def _build_line_fk(cls, bill_line_item, beneficiary):
-        bill_line_item["line_id"] = beneficiary.id
-        bill_line_item['line_type'] = ContentType.objects.get_for_model(beneficiary)
-
-    @classmethod
-    def _build_quantity(cls, bill_line_item):
-        bill_line_item["quantity"] = 1
-
-    @classmethod
-    def _build_price(cls, bill_line_item, amount):
-        bill_line_item["amount_total"] = amount
-        bill_line_item["unit_price"] = amount
+class BeneficiaryToBillItemConverter(BaseBenefitPackageStrategyItemConverter):
 
     @classmethod
     def _build_code(cls, bill_line_item, beneficiary):
         bill_line_item["code"] = f"{beneficiary.individual.first_name}-{beneficiary.individual.last_name}"
-
-    @classmethod
-    def _build_date_dates(cls, bill_line_item, payment_plan):
-        bill_line_item["date_valid_from"] = payment_plan.benefit_plan.date_valid_from
-        bill_line_item["date_valid_to"] = payment_plan.benefit_plan.date_valid_to
